cabinet/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
duckdb>=0.9.0
python-multipart>=0.0.6
email-validator>=2.1.0
pytest>=7.4.3
httpx>=0.25.0
markdown>=3.5.1
pyyaml>=6.0
markitdown>=0.1.1